Rename classes to get rid of "I" prefix for interfaces
