--- src/main/java/com/simprints/simface/embedding/IEmbeddingProcessor.kt
@@ -1,7 +1,0 @@
-package com.simprints.simface.embedding
-
-import android.graphics.Bitmap
-
-interface IEmbeddingProcessor {
-    fun getEmbedding(bitmap: Bitmap): ByteArray
-}
